pdf-service/requirements.txt
fastapi
uvicorn
python-multipart
openai
langchain
langchain-openai
langchain-community
langchain-classic
langchain-text-splitters
pymupdf
python-dotenv
pytesseract 